feat: session state machine, channel approval hardening, vault-based STT key resolution
- Add session state lifecycle (Created → Active → Running → Idle → Stopped/Error)
  with valid transition enforcement in session_manager ops
- Harden channel approval bridge: delivery timeouts, error sanitization,
  render-then-send pattern, expired approval cleanup
- Resolve STT provider API keys from vault at runtime instead of requiring
  them in config (api_key is #[serde(skip)])
- Expand exec security gate with additional validation
- Various small fixes across gateway handlers, extensions, and agent loop

+++ src/gateway/server/handler.rs
@@ -38,6 +38,7 @@ use super::per_client_buffer::PerClientBuffer;
 
 /// Shared context for handling a WebSocket connection.
 struct ConnectionContext {
+    #[allow(dead_code)]
     handlers: Arc<HandlerRegistry>,
     middleware_chain: MiddlewareChain,
     event_bus: Arc<GatewayEventBus>,
@@ -617,7 +618,7 @@ async fn handle_connection(
                         }
                     }
                     Err(broadcast::error::RecvError::Lagged(n)) => {
-                        buffer_metrics.add_overflow(n as u64);
+                        buffer_metrics.add_overflow(n);
                         debug!("Event forwarder lagged for {}, dropped {} events, total overflow={}", conn_id, n, buffer_metrics.overflow());
                         break;
                     }
